Rename DualSTA_SampleTool to MultipathLatencyAnalyzer
